Sprite Node explanation
Maybe adding a sprite on the screen could deserve its own tutorial/stub and link a reference to it. It will be important for artists who may not want to learn the entire engine but be on a team that uses the engine and needs to learn the quickest way possible to put in assets. A quick explanation on the sprite node in the tutorial could be helpful.